ADDED : SCC ALGORITHM
